backendpredict/requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
fastapi==0.103.2
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
idna==3.4
joblib==1.3.2
keras==2.14.0
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
ml-dtypes==0.2.0
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
Werkzeug==3.0.0
wrapt==1.14.1
